requirements.txt
Flask
SQLAlchemy
APScheduler
openpyxl
Werkzeug
flask_sqlalchemy
datetime
pandas
opencv-python-headless
face-recognition
ffmpeg-python